requirements.txt
langchain-core==1.0.2
langchain-groq==1.0.0
numpy==2.3.4
pandas==2.3.3
pandas-profiling==3.2.0
python-dotenv==1.2.1
RapidFuzz==3.14.3
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
tqdm==4.67.1
transformers==4.57.1
